docs : README.md 수정

+++ src/main/java/com/pplip/domain/trip/attraction/persistence/dao/SidoGugunsDao.java
@@ -5,6 +5,9 @@ import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
 
+/**
+ * 시도/구군 데이터에 접근하는 DAO 인터페이스
+ */
 @Mapper
 public interface SidoGugunsDao {
 
@@ -14,5 +17,5 @@ public interface SidoGugunsDao {
 
     List<AttractionResponse.Gugun> findAllGugun();
 
-    List<AttractionResponse.Gugun> findAllGugunInSido(Long sidoCode);
+    List<AttractionResponse.Gugun> findAllGugunInSido(Integer sidoCode);
 }
